feat: new features

+++ src/ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: dximenes <dximenes@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/05 18:25:43 by dximenes          #+#    #+#             */
-/*   Updated: 2025/04/09 16:30:02 by dximenes         ###   ########.fr       */
+/*   Updated: 2025/04/09 21:46:24 by dximenes         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
 	unsigned char	*str_dest;
 	unsigned char	*str_src;
-	int				i;
+	size_t			i;
 
 	str_dest = (unsigned char *)dest;
 	str_src = (unsigned char *)src;
